// src/shared/interfaces/permissions.ts
export enum Action {
  read = 'read',
  create = 'create',
  update = 'update',
  delete = 'delete',
  read_personal = 'read_personal',
  create_personal = 'create_personal',
  update_personal = 'update_personal',
  delete_personal = 'delete_personal',
}

export enum Resource {
  user = 'user',
  role = 'role',
  setting = 'setting',
  attendance = 'attendance',
}